tiered_church_formation_services/requirements.txt
## requirements.txt

# Django and Django REST framework
Django==3.2.19
djangorestframework==3.14.0

# Celery for asynchronous tasks
celery==5.2.7

# Stripe for payment processing
stripe==5.4.0

# PostgreSQL database adapter
psycopg2-binary==2.9.6

# Gunicorn for production server
gunicorn==20.1.0

# Redis for Celery backend and caching
redis==4.5.5

# Environment variable management
django-environ==0.10.0

# JWT authentication
djangorestframework-simplejwt==5.2.2

# CORS headers for API
django-cors-headers==4.0.0

# Pillow for image processing
Pillow==9.5.0

# Django extensions for additional tools
django-extensions==3.2.1

# Django debug toolbar for development
django-debug-toolbar==4.1.0

# Boto3 for AWS S3 integration (if used for file storage)
boto3==1.26.142

# Django Storages for additional storage backends
django-storages==1.13.2

# DRF-YASG for API documentation
drf-yasg==1.21.5

# Pytest for testing
pytest==7.3.1
pytest-django==4.5.2

# Coverage for test coverage reporting
coverage==7.2.7

# Flake8 for code linting
flake8==6.0.0

# Black for code formatting
black==23.3.0

# isort for import sorting
isort==5.12.0

# Django REST framework filters
django-filter==23.2

# Django REST framework JSON API
djangorestframework-jsonapi==6.0.0

# Django REST framework CSV renderer
djangorestframework-csv==2.1.1

# Django REST framework YAML renderer
djangorestframework-yaml==2.0.0

# Django REST framework XML renderer
djangorestframework-xml==2.0.0

# Django REST framework gzip
django-rest-framework-gzip==2.2.0

# Django REST framework camel case
djangorestframework-camel-case==1.4.2

# Django REST framework API key
djangorestframework-api-key==2.3.0

# Django REST framework recursive
djangorestframework-recursive==0.1.2

# Django REST framework bulk
djangorestframework-bulk==0.2.1

# Django REST framework datatables
djangorestframework-datatables==0.7.0

# Django REST framework word filter
djangorestframework-word-filter==0.2.0

# Django REST framework roles
djangorestframework-roles==0.3.0

# Django REST framework expiring tokens
djangorestframework-expiring-authtoken==0.1.4

# Django REST framework jsoneditor
django-rest-framework-jsoneditor==0.1.3

# Django REST framework social oauth2
django-rest-framework-social-oauth2==1.1.0

# Django REST framework encrypted lookup fields
djangorestframework-encrypted-lookup==1.0.1

# Django REST framework extensions
drf-extensions==0.7.1

# Django REST framework nested routers
drf-nested-routers==0.93.4

# Django REST framework generators
drf-generators==0.5.0

# Django REST framework encrypted fields
django-encrypted-model-fields==0.6.5

# Django REST framework recaptcha
django-recaptcha==3.0.0

# Django REST framework rate limit
django-ratelimit==4.0.0

# Django REST framework timed auth token
django-rest-knox==4.2.0

# Django REST framework gis
djangorestframework-gis==1.0

# Django REST framework guardian
djangorestframework-guardian==0.3.0

# Django REST framework tracking
drf-tracking==1.5.0

# Django REST framework passwordless
drfpasswordless==1.5.8

# Django REST framework channels
channels==4.0.0
channels-redis==4.1.0

# Django REST framework haystack
drf-haystack==1.8.11

# Django REST framework serializer extensions
drf-writable-nested==0.7.0

# Django REST framework dynamic fields
drf-dynamic-fields==0.3.1

# Django REST framework flex fields
drf-flex-fields==1.0.0

# Django REST framework spectacular
drf-spectacular==0.26.2

# Django REST framework nested serializers
drf-nested-serializer==0.1.1

# Django REST framework schema generator
drf-schema-adapter==3.0.4

# Django REST framework url filter
django-url-filter==0.3.15

# Django REST framework friendly errors
drf-friendly-errors==0.14

# Django REST framework sideloading
drf-sideloading==0.1.4

# Django REST framework proxy
django-rest-framework-proxy==1.6.0

# Django REST framework extra fields
drf-extra-fields==3.5.0

# Django REST framework compound fields
drf-compound-fields==2.0.0

# Django REST framework json field
django-jsonfield==1.4.1

# Django REST framework json field schema
drf-jsonfield-schema==0.2.0

# Django REST framework json schema
drf-json-schema==0.1.4

# Django REST framework json api
drf-json-api==0.11.0

# Django REST framework json api utils
drf-json-api-utils==1.0.3

# Django REST framework json api related field
drf-json-api-related-field==0.1.0

# Django REST framework json api serializer
drf-json-api-serializer==0.2.0

# Django REST framework json api filters
drf-json-api-filters==0.1.0
